Rename utils to config_loader
