+++ src/mantis/jira/__init__.py
@@ -2,4 +2,4 @@ from .jira_auth import JiraAuth
 from .jira_client import JiraClient
 from .jira_issues import JiraIssue, JiraIssues
 from .jira_options import JiraOptions, parse_args
-from .utils import JiraSystemConfigLoader
+from .config_loader import JiraSystemConfigLoader
