+++ src/mantis/jira/config_loader/__init__.py
@@ -1,0 +1,1 @@
+from .config_loader import JiraSystemConfigLoader
