--- src/mantis/jira/utils/__init__.py
@@ -1,1 +1,0 @@
-from .jira_system_config_loader import JiraSystemConfigLoader
